requirements.txt
Flask==3.1.2
numpy==1.26.4
Pillow==12.0.0
tensorflow==2.17.0
